+++ Attributes/RelationshipAttribute.php
@@ -52,6 +52,10 @@ abstract class RelationshipAttribute
     final public function init(Relationships $relationships): Relationship
     {
         $relationship = $this->addRelationship($relationships);
+        if (true === array_key_exists('orphanRemoval', $this->params)) {
+            $relationship->setOrphanRemoval($this->params['orphanRemoval']);
+        }
+
         $this->addClauses($relationship);
 
         return $relationship;